Fixes #2 - YAML with multiple records make the final result have lines duplicated bug
The code was iterating through all records for each line and appending the line each time it wasn't a match, so it was duplicating for each unmatched record.
It was created a new function "FindFirstRecordThatMatches" and used on main(). Given its result main() will decide if it will append the marked line or the original line.